phase_3/investor_bulletin/requirements.txt
# HTTP
fastapi
uvicorn
pydantic
requests

# Environment
python-dotenv

# Database
sqlalchemy
psycopg2-binary
sqlalchemy-cockroachdb
alembic

# Messageing
amqpstorm
pika

# Queuing and Scheduling
celery
# celery-beat
